+++ packages/components/src/private-apis.ts
@@ -12,6 +12,7 @@ import {
 	ValidatedTextareaControl,
 } from './validated-form-controls';
 import ContentEditableControl from './content-editable-control';
+import { useSlotFillsForNames } from './slot-fill/bubbles-virtually/use-slot-fills';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -31,6 +32,7 @@ lock( privateApis, {
 	},
 	Badge,
 	useDrag,
+	useSlotFillsForNames,
 	ValidatedInputControl,
 	ValidatedContentEditableControl,
 	ValidatedTextareaControl,